Combine some shared code with Yarn and QM Namespaces
